feat: add local font, change typo, logo
