+++ package.json
@@ -33,6 +33,7 @@
     "expo-font": "~11.0.1",
     "expo-local-authentication": "~13.0.2",
     "expo-secure-store": "~12.0.0",
+    "expo-splash-screen": "^0.17.5",
     "expo-status-bar": "~1.4.2",
     "lodash": "^4.17.21",
     "nativewind": "^2.0.11",
@@ -45,7 +46,7 @@
     "react-native-qrcode-svg": "^6.1.2",
     "react-native-safe-area-context": "4.4.1",
     "react-native-screens": "~3.18.0",
-    "react-native-svg": "^13.6.0",
+    "react-native-svg": "13.6.0",
     "react-native-vector-icons": "^9.2.0",
     "react-redux": "^8.0.5",
     "redux-thunk": "^2.4.2",
