+++ src/pages/Auth/EnablefaceId.tsx
@@ -3,8 +3,12 @@ import { View, Text } from 'react-native'
 import CButton from '../../components/basics/Button'
 import InterText from '../../components/basics/Button/InterText'
 import FullScreenContainer from '../../components/FullScreenContainer'
+import * as LocalAuthentication from 'expo-local-authentication';
 
 function EnablefaceId() {
+    const onPress = async () => {
+        await LocalAuthentication.authenticateAsync({})
+    }
     return (
         <FullScreenContainer passedClassName='flex-1  justify-between'>
             <View></View>
@@ -13,7 +17,7 @@ function EnablefaceId() {
                     <InterText passedClassName="text-xl pb-4" weight='100'>Privacy</InterText>
                     <InterText passedClassName="text-2xl pb-8" weight='600'>Enable biometric ID to secure your account</InterText>
                 </View>
-                <CButton passedClassName='w-full' theme='dark' onPress={() => { alert('Need apple device') }} >
+                <CButton passedClassName='w-full' theme='dark' onPress={onPress} >
                     Enable biometric ID
                 </CButton>
             </View>
