+++ src/pages/Auth/Welcome.tsx
@@ -1,4 +1,4 @@
-import { View, Text, ScrollView } from 'react-native'
+import { View, Text, Image, ScrollView } from 'react-native'
 import React, { useState } from 'react'
 import CButton from '../../components/basics/Button'
 import FullScreenContainer from '../../components/FullScreenContainer'
@@ -29,22 +29,15 @@ const Welcome = ({ navigation }) => {
     return (
         <FullScreenContainer passedClassName='flex-1  justify-between'>
             <ScrollView className="flex-auto">
-                <View className='h-32'>
-                    <Text>
-                        adsa
-                    </Text>
-                    <Text>
-                        adsa
-                    </Text>
-                    <Text>
-                        adsa
-                    </Text>
-                    <Text>
-                        adsa
-                    </Text>
+                <View className='h-64 justify-center'>
+                    <Image
+                        className=' w-12 '
+                        style={{resizeMode: "center"}}
+                        source={require('../../../assets/logo.png')}
+                    />
                 </View>
-                <View>
-                    <InterText passedClassName='text-4xl mt-6 font-semibold'>
+                <View className=''>
+                    <InterText passedClassName='text-4xl mt-6 font-semibold' weight='600'>
                         An anonymous wallet with institutional level security.
                     </InterText>
                 </View>
